refactor(storage): clean up iceberg report helper
